feat: setting the owner and other minor changes on the tasks crud

+++ backend/src/main/java/com/hahntask/backend/services/impl/TaskServiceImpl.java
@@ -2,6 +2,7 @@ package com.hahntask.backend.services.impl;
 
 import com.hahntask.backend.domain.entities.Task;
 import com.hahntask.backend.repositories.TaskRepository;
+import com.hahntask.backend.services.AuthService;
 import com.hahntask.backend.services.TaskService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,24 +15,37 @@ import java.util.UUID;
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository taskRepository;
+    private final AuthService authService;
 
     @Override
-    public List<Task> getAll() {
-        return taskRepository.findAll();
+    public List<Task> inbox() {
+        var auth = authService.getAuthenticatedUser();
+        return taskRepository.findByOwnerIdAndProjectId(auth.getId(), null);
     }
 
     @Override
     public Task find(UUID uuid) {
-        return taskRepository.findById(uuid).orElse(null);
+        var task = taskRepository.findById(uuid).orElseThrow();
+        var auth = authService.getAuthenticatedUser();
+        if (!auth.getId().equals(task.getOwner().getId())) {
+            return null;
+        }
+        return task;
     }
 
     public Task create(Task task) {
+        var auth = authService.getAuthenticatedUser();
+        task.setOwner(auth);
         return taskRepository.save(task);
     }
 
     @Override
     public Task update(UUID uuid, Task task) {
         var existing = taskRepository.findById(uuid).orElseThrow();
+        var auth = authService.getAuthenticatedUser();
+        if (!auth.getId().equals(existing.getOwner().getId())) {
+            return null;
+        }
         existing.setTitle(task.getTitle());
         existing.setDescription(task.getDescription());
         existing.setPriority(task.getPriority());
@@ -41,6 +55,11 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public void delete(UUID uuid) {
+        var task = taskRepository.findById(uuid).orElseThrow();
+        var auth = authService.getAuthenticatedUser();
+        if (!auth.getId().equals(task.getOwner().getId())) {
+            return;
+        }
         taskRepository.deleteById(uuid);
     }
 }
